+++ src/components/PostOptionsModal/PostOptionsModal.jsx
@@ -1,19 +1,18 @@
 import { useEffect, useState } from "react";
-import { useAuth } from "../../contexts/auth-context";
-import { usePosts } from "../../contexts/post-context";
-import { useUsers } from "../../contexts/user-context";
+import { useLocation, useNavigate } from "react-router-dom";
+import { useAuth, usePosts, useUsers } from "../../index";
+import { PostModal } from "..";
 import {
   FaEdit,
   FaTrash,
   FaUserPlus,
   RiUserUnfollowFill,
 } from "../../utils/icons";
-import { PostModal } from "..";
-import { useLocation, useNavigate } from "react-router-dom";
 import { Modal } from "@mui/material";
 
 const PostOptionsModal = ({ post, setShowOptions }) => {
   const { _id, username } = post;
+
   const { currentUser } = useAuth();
   const {
     usersState: { users },
