+++ src/components/UsersModal/UsersModal.jsx
@@ -1,22 +1,12 @@
 import { useNavigate } from "react-router-dom";
+import { UserAvatar } from "..";
+import { styles } from "../../utils/constants";
 import { FaTimes } from "../../utils/icons";
-import { UserAvatar } from "../UserAvatar/UserAvatar";
 
 const UsersModal = ({ usersListModal, setUsersListModal }) => {
   const { title, list } = usersListModal;
 
   const navigate = useNavigate();
-
-  const styles = {
-    position: "absolute",
-    top: "50%",
-    left: "50%",
-    transform: "translate(-50%, -50%)",
-    bgcolor: "background.paper",
-    border: "1px solid #000",
-    boxShadow: 24,
-    p: 4,
-  };
 
   return (
     <div
