+++ src/utils/constants.js
@@ -30,6 +30,17 @@ export const guestUserLoginDetails = {
   password: "emily@123Smith",
 };
 
+export const styles = {
+  position: "absolute",
+  top: "50%",
+  left: "50%",
+  transform: "translate(-50%, -50%)",
+  bgcolor: "background.paper",
+  border: "1px solid #000",
+  boxShadow: 24,
+  p: 4,
+};
+
 export const actionTypes = {
   GET_ALL_POSTS: "GET_ALL_POSTS",
   LIKE_POST: "LIKE_POST",
